feat: Thumbnailing - backend implementation (#155)
* fix: change module resolution to support new libraries

* refactor: split the many functions in multiple files

* style: fix eslint and update package-lock

* feat(backend): add seed data for firebase

* feat: implement generateThumbnail

* feat: add the onCall variant of the generateThumbnail

* feat: crop using entropy using AU

* style: replace magic numbers with named constants
