+++ cloud-functions/functions/.eslintrc.cjs
@@ -29,5 +29,6 @@ module.exports = {
     "quotes": ["error", "double"],
     "import/no-unresolved": 0,
     "indent": ["error", 2],
+    "max-len": "off",
   },
 };
