+++ cloud-functions/functions/firebase-seed/firebase-export-metadata.json
@@ -1,0 +1,12 @@
+{
+  "version": "13.24.1",
+  "firestore": {
+    "version": "1.19.8",
+    "path": "firestore_export",
+    "metadata_file": "firestore_export/firestore_export.overall_export_metadata"
+  },
+  "storage": {
+    "version": "13.24.1",
+    "path": "storage_export"
+  }
+}
