+++ cloud-functions/functions/firebase-seed/storage_export/buckets.json
@@ -1,0 +1,7 @@
+{
+  "buckets": [
+    {
+      "id": "travelpouch-7d692.appspot.com"
+    }
+  ]
+}
