+++ cloud-functions/functions/firebase-seed/storage_export/metadata/d94246d9-355b-4143-947a-06abc13ff759.json
@@ -1,0 +1,19 @@
+{
+  "name": "myDocument2",
+  "bucket": "travelpouch-7d692.appspot.com",
+  "metageneration": 1,
+  "generation": 1731638646088,
+  "contentType": "application/octet-stream",
+  "storageClass": "STANDARD",
+  "contentDisposition": "inline",
+  "downloadTokens": [
+    "35169ab0-db81-4b47-a36f-ae0b77812cd4"
+  ],
+  "etag": "z6y7HEy4AxmRHfzGZnR+b9Cl66s",
+  "customMetadata": {},
+  "timeCreated": "2024-11-15T02:44:06.088Z",
+  "updated": "2024-11-15T02:44:06.088Z",
+  "size": 1165031,
+  "md5Hash": "AotkaIrYW6ShfSjzrSHFTA==",
+  "crc32c": "654433733"
+}
