+++ cloud-functions/functions/package.json
@@ -16,7 +16,9 @@
   "main": "lib/index.js",
   "dependencies": {
     "firebase-admin": "^12.1.0",
-    "firebase-functions": "^6.0.1"
+    "firebase-functions": "^6.0.1",
+    "mupdf": "^0.3.0",
+    "wasm-vips": "^0.0.11"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^5.12.0",
@@ -27,5 +29,6 @@
     "firebase-functions-test": "^3.1.0",
     "typescript": "^4.9.0"
   },
-  "private": true
+  "private": true,
+  "type": "module"
 }
