+++ cloud-functions/functions/tsconfig.dev.json
@@ -1,5 +1,5 @@
 {
   "include": [
-    ".eslintrc.js"
+    ".eslintrc.cjs"
   ]
 }
